receive: drop packets in low memory conditions instead of using reserves
The crypto needs atomic vmap anyway, which means if packets make it in
a memalloc situation, they might not make it much further. We can
revisit memalloc semantics later if people are stacking this with NFS or
the like, and also swapping to NFS, but now it's a bit overkill and
introduces undue complications.